+++ quantum/quantum.c
@@ -24,8 +24,10 @@
 #    include "outputselect.h"
 #endif
 
-#include "backlight.h"
-extern backlight_config_t backlight_config;
+#ifdef BACKLIGHT_ENABLE
+#    include "backlight.h"
+    extern backlight_config_t backlight_config;
+#endif
 
 #ifdef FAUXCLICKY_ENABLE
 #    include "fauxclicky.h"
