+++ tmk_core/common/command.c
@@ -32,9 +32,12 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #include "sleep_led.h"
 #include "led.h"
 #include "command.h"
-#include "backlight.h"
 #include "quantum.h"
 #include "version.h"
+
+#ifdef BACKLIGHT_ENABLE
+#    include "backlight.h"
+#endif
 
 #ifdef MOUSEKEY_ENABLE
 #    include "mousekey.h"
